+++ Assets/General/Architecture/Scripts/Services/EventService.cs
@@ -15,7 +15,7 @@ namespace General.Services
         }
         
         
-        public void SpawnWarrior(BattleSide battleSide, WarriorType type, bool isBoss, byte squadID, Transform spawnPoint)
+        public void SpawnWarrior(BattleSide battleSide, WarriorType type, bool isBoss, int squadID, Transform spawnPoint)
         {
             _world.NewEntity().Get<SpawnWarriorEvent>() = new SpawnWarriorEvent
             {
